+++ Assets/Colonization/Scripts/Characteristic/Skills/AttackSkill.cs
@@ -12,7 +12,7 @@ namespace Vurbiri.Colonization
         {
             public AnimationClipSettingsScriptable clipSettings;
             public bool isValid;
-            public float percentDamage;
+            public int percentDamage;
             public AttackState.Settings settings;
             public AttackSkillUI ui;
         }
